pyproject.toml
[tool.poetry]
name = "flask_app"
version = "0.1.0"
description = "Flask API backend"
authors = ["Danil Chekh <chekhwork@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.0.3"
flask-cors = "^5.0.0"
flask-sqlalchemy = "^3.1.1"
flasgger = "^0.9.7.1"
alembic = "^1.14.0"
psycopg2-binary = "^2.9.10"
flask-jwt-extended = "^4.7.1"
gunicorn = "^23.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-cov = "^6.0.0"
python-dotenv = "^1.0.1"

[tool.pytest.ini_options]
# Coverage settings
addopts = "--cov=app --cov=config --cov=models --cov=routes"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
